fix: Ensure image URLs quoted before download
When trying to use pypub to generate an ebook, the rendering process
failed when it tried to download the image locate at
https://teatimechinese.com/wp-content/uploads/2022/05/黄巾.jpg.

That happens because the Chinese characters need to be URL encoded prior
to the request being made.

This change makes sure that image URLs get properly encoded before
sending the GET request to download them.